% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdoc.R
\name{gtemplate}
\alias{gtemplate}
\title{Get an R Markdown document for practice}
\usage{
gtemplate(filename = "test.Rmd")
}
\arguments{
\item{filename}{filename}
}
\description{
Get an R Markdown document for practice
}
\examples{
{
\dontrun{
my_rmd <- gtemplate()
rmarkdown::render(my_rmd)
}
}
}

