% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdoc.R
\name{gdoc}
\alias{gdoc}
\title{Create a Google Doc from an R Markdown File}
\usage{
gdoc(template = NULL, token = gd_auth(), keep_md = FALSE,
  clean_supporting = TRUE)
}
\arguments{
\item{template}{An MS word file with styles for your document.  NULL uses
package defaults}

\item{token}{Google Auth token. \code{\link{gd_auth}} will generate one by
default}

\item{keep_md}{Boolean.  Whether to retain the knit markdown file}

\item{clean_supporting}{Boolean. Whether to keep other intermediate files,
including the MS Word file.}
}
\description{
Create a Google Doc from an R Markdown File
}

